プリロードを分離
